+++ stubs/livewire/resources/views/components/connected-account.blade.php
@@ -35,7 +35,7 @@
 
                 @if (! empty($createdAt))
                     <div class="text-xs text-gray-500">
-                        Connected on {{ $createdAt }}
+                        Connected {{ $createdAt }}
                     </div>
                 @else
                     <div class="text-xs text-gray-500">